pyproject.toml
[tool.poetry]
name = "google_cloud_logger"
version = "0.1.0"
description = ""
authors = ["Sam Yao <turisesonia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-logging = "^1.15.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
